+++ zict/__init__.py
@@ -1,3 +1,4 @@
+from zict.async_buffer import AsyncBuffer as AsyncBuffer
 from zict.buffer import Buffer as Buffer
 from zict.cache import Cache as Cache
 from zict.cache import WeakValueMapping as WeakValueMapping